Refactor barrier controller to split out edit functionality
